# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/login": {
      "post": {
        "description": "Otorga acceso a un usuario registrado",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Datos de usuario",
            "in": "body",
            "description": "Datos JSON para la api",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " 'login':True "
          },
          "400": {
            "description": "Credenciales Invalidas"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "description": "Cierre de sesion del usuario logeado",
        "produces": [
          "application/json"
        ],          
        "responses": {
          "200": {
            "description": " 'logout': True"
          }
        }
      }
    },
    "/api/order/list": {
      "get": {
        "description": "Filtra órdenes basadas en: material_id o material_name (Si hay id, este se desestima), rango de fechas, y si las órdenes no están reservadas y aún están disponibles.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "query",
            "description": "ID del material para filtrar",
            "required": false,
            "type": "integer"
          },
          {
            "name": "material_name",
            "in": "query",
            "description": "Nombre del material para filtrar",
            "required": false,
            "type": "string"
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Fecha de inicio del rango (formato: YYYY-MM-DD)",
            "required": false,
            "type": "string"
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Fecha de fin del rango (formato: YYYY-MM-DD)",
            "required": false,
            "type": "string"
          },
          {
            "name": "only_unreserved",
            "in": "query",
            "description": "Si es true, solo muestra órdenes no reservadas y disponibles",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve una lista de órdenes filtradas"
          },
          "404": {
            "description": "No se encontraron órdenes con los filtros proporcionados"
          }
        }
      }
    },
    "/api/order/{order_id}/delivered": {
      "put": {
        "description": "Marca una orden como entregada si no lo está ya.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID de la orden",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "La orden fue marcada como entregada correctamente"
          },
          "400": {
            "description": "La orden no existe o ya fue entregada"
          },
          "403": {
            "description": "No tienes permisos para entregar la orden o no se puede entregar una orden no reservada"
          }
        }
      }
    },
    "/api/order/{order_id}/reserve": {
      "post": {
        "description": "Asigna un proveedor a una orden si no ha sido reservada ya y el proveedor está registrado para el material de la orden.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID de la orden a reservar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "La orden fue reservada correctamente"
          },
          "400": {
            "description": "La orden ya ha sido reservada o ya no se encuentra disponible"
          },
          "403": {
            "description": "No tienes permisos para reservar la orden o el proveedor no está registrado para este material"
          },
          "404": {
            "description": "La orden no existe"
          }
        }
      }
    },
    "/api/order/material/register/{material_id}": {
      "post": {
        "description": "Registra un proveedor para un material dado",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "description": "ID del material al cual registrarse",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Se registró correctamente para el material especificado"
          },
          "403": {
            "description": "No tienes permisos para reservar la orden"
          },
          "404": {
            "description": "No se encontró el material"
          }
        }
      }
    },
    "/api/order/material_list": {
      "get": {
        "description": "Muestra la lista de materias de ids con sus nombres",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de materiales"
          }
        }
      }
    }
  ,
    "/api/order/myorders": {
      "get": {
        "description": "Devuelve las ordenes que el proveedor tiene a su cargo. Se pueden filtrar si solo estan en estado reservadas o si ya fueron entregadas.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Adminte 2 valores 'reserved' y 'delivered'",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve una lista de órdenes asignadas al usuario"
          },
          "404": {
            "description": "No se encontraron órdenes asignadas"
          }
        }
      }
    },
    "/api/db/reset": {
      "post": {
        "description": "Resetear la base de datos. Elimina las tuplas y crea las tablas.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Database reset successful"
          },
          "400": {
            "description": "Database reset failed"
          }
        }
      }
    },
    "/api/db/seeds": {
      "post": {
        "description": "Reepoblar la base de datos con datos de prueba.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Seeding database successful"
          },
          "400": {
            "description": "Database seeding failed"
          }
        }
      }
    }
  },
  "definitions": {
    "LoginData": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}